5.19 delwarcbyid已经完成，优化问题正在进行中 detail页高亮等已完成，bug基本修复，session问题待解决

+++ src/main/java/com/bishe/bishe/es/EsDaoImpl.java
@@ -81,18 +81,18 @@ public class EsDaoImpl implements EsDao {
     public Boolean deleteDocument(String index, String type, String id) {
         Delete delete = new Delete.Builder(id).index(index).type(type).build();
         JestResult result = null ;
-        Integer successful = null;
+        Double successful = null;
         try {
             result = client.execute(delete);
             JsonObject jsonObject = result.getJsonObject();
             Gson gson = new Gson();
             Map tempmap = gson.fromJson(jsonObject.getAsJsonObject("_shards"),Map.class);
-            successful = (Integer) tempmap.get("successful");
+            successful = (Double) tempmap.get("successful");
             log.info("deleteDocument == " + result.getJsonString());
         } catch (Exception e) {
             e.printStackTrace();
         }
-        return successful!=0;
+        return successful!=(Double) 0.0;
     }
 
     /**
